[BandcampBridge] Add option for modified albums
Albums with new tracks added to them can now optionally show up as new
articles. This however means each album page has to be parsed, so an
option to limit the number of albums to load was also added. This also
means only the albums which are parsed will be checked for new tracks.